+++ src/main/kotlin/com/dobby/backend/application/mapper/MemberMapper.kt
@@ -1,6 +1,6 @@
 package com.dobby.backend.application.mapper
 
-import com.dobby.backend.application.usecase.member.GetMyExperimentPostsUseCase
+import com.dobby.backend.application.usecase.experiment.GetMyExperimentPostsUseCase
 import com.dobby.backend.domain.model.experiment.Pagination
 
 object MemberMapper {
