+++ src/main/kotlin/com/dobby/backend/application/service/ExperimentPostService.kt
@@ -5,11 +5,14 @@ import com.dobby.backend.application.usecase.experiment.*
 import com.dobby.backend.application.usecase.experiment.CreateExperimentPostUseCase
 import com.dobby.backend.application.usecase.experiment.GetExperimentPostApplyMethodUseCase
 import com.dobby.backend.application.usecase.experiment.GetExperimentPostDetailUseCase
+import com.dobby.backend.application.usecase.experiment.GetMyExperimentPostTotalCountUseCase
+import com.dobby.backend.application.usecase.experiment.GetMyExperimentPostsUseCase
 import com.dobby.backend.domain.exception.ExperimentAreaInCorrectException
 import com.dobby.backend.domain.exception.ExperimentAreaOverflowException
 import com.dobby.backend.infrastructure.database.entity.enums.areaInfo.Area
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
+import java.security.InvalidParameterException
 
 @Service
 class ExperimentPostService(
@@ -22,7 +25,9 @@ class ExperimentPostService(
     private val generateExperimentPostPreSignedUrlUseCase: GenerateExperimentPostPreSignedUrlUseCase,
     private val updateExpiredExperimentPostUseCase: UpdateExpiredExperimentPostUseCase,
     private val updateMyExperimentPostRecruitStatusUseCase: UpdateMyExperimentPostRecruitStatusUseCase,
-    private val getExperimentPostTotalCountByCustomFilterUseCase: GetExperimentPostTotalCountByCustomFilterUseCase
+    private val getExperimentPostTotalCountByCustomFilterUseCase: GetExperimentPostTotalCountByCustomFilterUseCase,
+    private val getMyExperimentPostsUseCase: GetMyExperimentPostsUseCase,
+    private val getMyExperimentPostTotalCountUseCase: GetMyExperimentPostTotalCountUseCase,
 ) {
     @Transactional
     fun createNewExperimentPost(input: CreateExperimentPostUseCase.Input): CreateExperimentPostUseCase.Output {
@@ -88,5 +93,22 @@ class ExperimentPostService(
 
     fun getExperimentPostTotalCount(input: GetExperimentPostTotalCountByCustomFilterUseCase.Input): Int {
         return getExperimentPostTotalCountByCustomFilterUseCase.execute(input)
+    }
+
+    @Transactional
+    fun getMyExperimentPosts(input: GetMyExperimentPostsUseCase.Input): List<GetMyExperimentPostsUseCase.Output> {
+        validateSortOrder(input.pagination.order)
+        return getMyExperimentPostsUseCase.execute(input)
+    }
+
+    fun getMyExperimentPostsCount(input: GetMyExperimentPostTotalCountUseCase.Input): GetMyExperimentPostTotalCountUseCase.Output {
+        return getMyExperimentPostTotalCountUseCase.execute(GetMyExperimentPostTotalCountUseCase.Input(input.memberId))
+    }
+
+    private fun validateSortOrder(sortOrder: String): String {
+        return when (sortOrder) {
+            "ASC", "DESC" -> sortOrder
+            else -> throw InvalidParameterException("Invalid sort order. Please use 'ASC' or 'DESC'")
+        }
     }
 }
