+++ src/main/kotlin/com/dobby/backend/application/service/MemberService.kt
@@ -1,14 +1,13 @@
 package com.dobby.backend.application.service
 
-import com.dobby.backend.application.usecase.experiment.UpdateMyExperimentPostRecruitStatusUseCase
-import com.dobby.backend.application.usecase.member.GetMyExperimentPostTotalCountUseCase
+import com.dobby.backend.application.usecase.experiment.GetMyExperimentPostTotalCountUseCase
+import com.dobby.backend.application.usecase.experiment.GetMyExperimentPostsUseCase
 import com.dobby.backend.application.usecase.member.*
 import com.dobby.backend.domain.exception.SignupOauthEmailDuplicateException
 import com.dobby.backend.domain.gateway.member.MemberGateway
 import com.dobby.backend.infrastructure.database.entity.enums.MemberStatus
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
-import java.security.InvalidParameterException
 
 @Service
 class MemberService(
@@ -18,8 +17,6 @@ class MemberService(
     private val verifyResearcherEmailUseCase: VerifyResearcherEmailUseCase,
     private val getResearcherInfoUseCase: GetResearcherInfoUseCase,
     private val getParticipantInfoUseCase: GetParticipantInfoUseCase,
-    private val getMyExperimentPostsUseCase: GetMyExperimentPostsUseCase,
-    private val getMyExperimentPostTotalCountUseCase: GetMyExperimentPostTotalCountUseCase,
 ) {
     @Transactional
     fun participantSignup(input: CreateParticipantUseCase.Input): CreateParticipantUseCase.Output {
@@ -42,22 +39,5 @@ class MemberService(
     @Transactional
     fun getParticipantInfo(input: GetParticipantInfoUseCase.Input): GetParticipantInfoUseCase.Output {
         return getParticipantInfoUseCase.execute(input)
-    }
-
-    @Transactional
-    fun getMyExperimentPosts(input: GetMyExperimentPostsUseCase.Input): List<GetMyExperimentPostsUseCase.Output> {
-        validateSortOrder(input.pagination.order)
-        return getMyExperimentPostsUseCase.execute(input)
-    }
-
-    private fun validateSortOrder(sortOrder: String): String {
-        return when (sortOrder) {
-            "ASC", "DESC" -> sortOrder
-            else -> throw InvalidParameterException("Invalid sort order. Please use 'ASC' or 'DESC'")
-        }
-    }
-
-    fun getMyExperimentPostsCount(input: GetMyExperimentPostTotalCountUseCase.Input): GetMyExperimentPostTotalCountUseCase.Output {
-        return getMyExperimentPostTotalCountUseCase.execute(GetMyExperimentPostTotalCountUseCase.Input(input.memberId))
     }
 }
